+++ src/ngraph/runtime/interpreter/CMakeLists.txt
@@ -15,7 +15,7 @@
 # ******************************************************************************
 
 if (NGRAPH_INTERPRETER_ENABLE)
-    add_library(interpreter_backend SHARED int_backend.cpp node_wrapper.cpp)
+    add_library(interpreter_backend SHARED int_backend.cpp node_wrapper.cpp int_executable.cpp)
     if(NGRAPH_LIB_VERSIONING_ENABLE)
         set_target_properties(interpreter_backend PROPERTIES
             VERSION ${NGRAPH_VERSION}
